Remove commented out code
If you have your code under version control (like git), you probably don't need
old or experimental code to hang around in comments: old code was (ideally)
already committed in the past and it's better to create a separate branch for
experimental code.

Without unnecessary comments it's easier to follow the actual code for someone
unfamiliar with it.

More info:
- https://try.github.io/ — quick interactive introduction to Git
- https://git-scm.com/book/en/v2 — a more thorough introduction
- https://www.sourcetreeapp.com/ — if you prefer to use GUI

Mostly GitHub specific, but sometimes applicable to Git in general:
- https://help.github.com/articles/tracing-changes-in-a-file/
- https://help.github.com/articles/comparing-commits-across-time/
- https://help.github.com/articles/differences-between-commit-views/